Documentation: full-ness of a bundle is significant for cloning
Not necessarily every bundle file can be cloned from.  Only the ones
that do not need prerequisites can.

When 50cfbeb (Documentation: minor grammatical fixes and rewording
in git-bundle.txt, 2009-03-22) reworded this paragraph, it lost a
critical hint to tell readers why this particular bundle can be
cloned from.  Resurrect it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>